Split the test
